config.json: Fix check-layer for meta-poky/meta-yocto-bsp
Currently, the tests for meta-poky and meta-yocto-bsp were malfunctioning.
We need to a) set a distro other than poky and b) avoid the isolation wrapper
since it will use the default environment and reset to the poky distro and
default bblayers.conf which will disable the layer test.

With those two fixes, the layer checks work correctly.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>