config.json: set ROOT_HOME to /root for sysvinit with systemd
For sysvinit with systemd, the systemd package will still be build.
This gives us a warning like below:

  WARNING: systemd-1_255.1-r0 do_install: Using /home/root as root
           user's home directory is not fully supported by systemd

Set ROOT_HOME to "/root" to avoid such warning.

Note that when using sysvinit as the init manager, /home/root is totally
valid, it's just that 'systemd' being in DISTRO_FEATURES causes systemd
to be built. So the only purpose of this patch's is to avoid warnings
on autobuilders.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>