config.json: run CDN population tasks with --runall build
There was a report of missing go-cross:populate_sysroot here:
https://autobuilder.yoctoproject.org/typhoon/#/builders/83/builds/6504/steps/23/logs/stdio

The builds that are expected to create the object (or take it from cache) are:
https://autobuilder.yoctoproject.org/typhoon/#/builders/42/builds/8481/steps/12/logs/stdio
https://autobuilder.yoctoproject.org/typhoon/#/builders/76/builds/8457/steps/25/logs/stdio
and neither mentioned that task anywhere, instead going directly to go-helloworld's rpm.

This means that go-cross can be neither taken from cache not written into it in
a-full, and we need '--runall build' to ensure it's not skipped over.

I also took the opportunity to match up the target image names with what the CDN
test actually checks.

Signed-off-by: Alexander Kanavin <alex.kanavin@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>