config.json: Switch poky-altcfg to use ipk as the package backend
Since we plan to have better "binary distro" testing on the autobuilder
and we plan to primarily use ipk for this, having ipk based images available
as part of our builds is desireable.

Since we want systemd based images with ipk, it makes sense to leave standard
poky using rpm images but switch poky-altcfg to use ipk.

Whilst late in the release cycle, doing this now for scarthgap gives us a good
platform to build off in the long term too.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>