config.json: add sbsa-ref builds to meta-arm
This machine replaces qemu-generic-arm64, which was built on the AB until
60825ac6 landed.

Because the sbsa-ref machine boots a disk image inside qemu there is no
way for the static IPs to be passed to the kernel directly, which means
TAP networking fails. Work around this by always running the tests with
slirp networking.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>