config.json: add a builder for meta-webosose
Add a minimal (although still substantial) build of meta-webosose for LG.
Because meta-webosose has complex layer appends and dependencies we need
to manually set BBFILE_PRIORITY and rewrite bblayer.conf to match what
their custom build scripts currently do.

Builds against oe-core master are known to not work at present, but this
will be backported to kirkstone which is the latest supported release.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>