bugtriage: consider more point releases as in the future
There is a bug which has the Target Milestone of 5.0.15 and appears in
the Old Milestone list despite the latest 5.0.x release being 5.0.10.

This happened because we only considered the next three point releases,
so extend the list of future point releases to ten to ensure that we
include all realistic versions.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>