scripts/yocto-supported-distros: check worker status on AB API
Add a --check-worker-statuses flag to also check the workers statuses on
the Autobuilder, and filter out any worker that is currently
disconnected.

This should help maintaining the config.py file in the Autobuilder. This
is relevant for the master branch here as we do not need to maintain
poky.conf as strictly on stable branches.

This should warn the following on master, at the time of writing:

  WARNING: Missing on the autobuilder but listed in poky.conf: ubuntu-20.04

Meaning ubuntu-20.04 should be removed from poky.conf on meta-poky's
master.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>