scripts/yocto-supported-distros: add option to print old distros
We can use config.py to list distributions that used to be tested for a
given release, by comparing the values in workers_prev_releases (all
workers for a release) and all_workers (all active workers for any
release).

Add the --old-distros to do that. Make it mutually exclusive to
--compare, we don't want to do multiple things at the same time.

Add a pattern to match centos, as it was defined for kirkstone in the
past (we use stream now).

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>