scripts: Add bitbake-setup repo filtering for migration
In order to allow bitbake-setup builds to co-exist with older poky ones,
we need to add the bitbake repository to the layers present in the builds.

For older branches, add a workaround to filter the layerinfo.json file so
the right information is presented in older builds but newer ones can
work for bitbake-setup too.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>