config.json/scripts: Switch to use bitbake-setup
With the addition of bitbake-setup to the master branch, we can switch from using the
combo-layer built poky repository to using bitbake-setup instead.

This patch set changes the setup to use bitbake/oecore/meta-yocto/yocto-docs instead
of poky. Most of the changes are due to the new layer layout, or using bitbake-setuip
instead of scripts in helper. The existing scripts are still run for now to handle layers
not in the core configuration json.

We continue to write extra settings to auto.conf.

This switches the code to use the distro/machine selection from fragments.

We need to use init-build-env instead of oe-init-build-env in most cases.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>