scripts/getproperties: Update to return all revisions/branches and compute build_revision
Now that bitbake-setup is fetching the layers, we need to report back the revisions
fetched back to buildbot. We can do this by sharing extra properties back
to the controller from this script.

The buildbot UI needs a way to group the "same" builds together. Compute a dummy
sha1 hash of the bitbake/oecore/helper revisions to do that which can be used in
place of the poky hash there.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>