scripts/getproperties: Fix to handled detached head states
Now that we have a whinlatter branch, we're seeing HEAD being reported instead
of the real branch name. For some reason buildbot leaves the tree in a detached
head state even though the revisions match and this doesn't work with rev-parse.

Try using name-rev instead (thanks Mathieu for the pointer).

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>