scripts/run-config: Use correct branchname/repo after bitbake-setup changes
The controller is currently passing in information about poky which is incorrect.
Ideally we swap this for the right information but for compatibility with older
release builds that is trickier. The plan is to ultimately use the properties
json files to transfer this in future so shortcut to that a little and hack
that into place now to fix the performance testing builds.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>